Remove the debconf to allows user inputs during the upgrade, update or install packages
The pengwin setup keep waiting for a user input during the upgrade process. The problem was when the upgrade (or update) was executed with  debconf-apt-progress, any input required to complete the upgrade (or update) was not displaying to the user.

Now the upgrade is execute without a graphical progress bar.

Implement : PEN-124 .