Make Poetry available in current shell after installation
Ran into this when starting a fresh new Pengwin install. Found this issue and comment from Poetry: https://github.com/python-poetry/poetry/issues/532#issuecomment-431417497 which suggests sourcing it. Also without sourcing, restarting WSL doesn't cause the shell to update, which is odd. This change fixes the issue presuming it's the normal user shell  that's sources and not root. It will fail in the case of root unless pengwin setup was run directly by root (which is not supported by Pengwin).